+++ src/Simpletools/Cloud/Google/Storage/Client.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Simpletools\Cloud\Google\Storage;
+
+class Client
+{
+    protected static $_client;
+    public static function get()
+    {
+        if(!self::$_client)
+        {
+            self::$_client  = \Simpletools\Cloud\Google\Client::get()->storage();
+        }
+
+        return self::$_client;
+    }
+}